Creates package for the Signaling Server API public elements.
- Version, with prefix `v1`, the current services endpoints.
